milvus-test-python/pyproject.toml
[project]
name = "milvus-test-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langchain-community>=0.3.0",
    "dashscope>=1.20.0",
    "pymilvus>=2.6.10",
    "python-dotenv>=1.2.2",
    "ebooklib>=0.18",
    "html2text>=2024.2.26",
]
